Neutron driver: split classes into separate modules
