+++ calico/openstack/t_etcd.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2015 Metaswitch Networks
+# All Rights Reserved.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
+# Etcd-based transport for the Calico/OpenStack Plugin.
+
+# Calico imports.
+from calico.openstack.transport import CalicoTransport
+
+LOG = None
+
+
+class CalicoTransportEtcd(CalicoTransport):
+    """Calico transport implementation based on etcd."""
+
+    def __init__(self, driver, logger):
+        super(CalicoTransportEtcd, self).__init__(driver)
+
+        # Initialize logger.
+        global LOG
+        LOG = logger
+
+    def initialize(self):
+        pass
+
+    def endpoint_created(self, port):
+        pass
+
+    def endpoint_updated(self, port):
+        pass
+
+    def endpoint_deleted(self, port):
+        pass
+
+    def security_group_updated(self, sg):
+        pass
